Create new transform test file & add test for custom preset feature
