+++ azure-cli2017.pyproj
@@ -140,6 +140,9 @@
     <Compile Include="command_modules\azure-cli-advisor\azure_bdist_wheel.py" />
     <Compile Include="command_modules\azure-cli-advisor\setup.py" />
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\commands.py" />
+    <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\tests\latest\test_ams_transform_scenarios.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="command_modules\azure-cli-ams\azure\cli\command_modules\ams\_sdk_utils.py">
       <SubType>Code</SubType>
     </Compile>
