+++ src/command_modules/azure-cli-ams/azure/cli/command_modules/ams/tests/latest/data/invalidCustomPreset.json
@@ -1,32 +1,11 @@
-erval": "00:00:02",
-        "H264Layers": [
-          {
-            "Profile": "Auto",
-            "Level": "auto",
-            "Bitrate": 1045,
-            "MaxBitrate": 1045,
-            "BufferWindow": "00:00:05",
-            "ReferenceFrames": 3,
-            "EntropyMode": "Cavlc",
-            "AdaptiveBFrame": true,
-            "Type": "H264Layer",
-            "Width": "640",
-            "Height": "360",
-            "FrameRate": "0/1"
-          }
-        ],
-        "Type": "H264Video"
-      },
-      {
-        "Type": "CopyAudio"
+{
+  "Version": 1,
+  "Outputs": [
+    {
+      "FileName": "{Basename}{Extension}",
+      "Format": {
+        "Type": "MP4Format"
       }
-    ],
-    "Outputs": [
-      {
-        "FileName": "{Basename}{Extension}",
-        "Format": {
-          "Type": "MP4Format"
-        }
-      }
-    ]
-  }
+    }
+  ]
+}
